refactor ui components: testSuiteImport & login dialog
